requirements.txt
# Core Dependencies
python-binance>=1.0.17
python-dotenv>=1.0.0
requests>=2.31.0

# Testing Dependencies
pytest>=7.4.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0